ups, need to wrap expression from last commit in ( … ) ref:00322c5169e08a9deacf05caceee83c3f2edc808

+++ catboost/cuda/data/load_data.cpp
@@ -159,7 +159,7 @@ namespace NCatboostCuda {
                                                   nanMode,
                                                   borders);
 
-                const int binCount = borders.size() + 1 + ENanMode::Forbidden != nanMode;
+                const int binCount = static_cast<const int>(borders.size() + 1 + (ENanMode::Forbidden != nanMode));
                 auto compressedLine = CompressVector<ui64>(binarizedData, IntLog2(binCount));
 
                 featureColumns[featureId] = MakeHolder<TBinarizedFloatValuesHolder>(featureId,
